// src/lib/api/generated/admin/models/memberDtoRequest.ts
/**
 * Generated by orval v7.5.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { MemberDtoRequestRole } from './memberDtoRequestRole';

export interface MemberDtoRequest {
  companyId?: number;
  loginId?: string;
  password?: string;
  name?: string;
  email?: string;
  role?: MemberDtoRequestRole;
  phoneNumber?: string;
  birthDate?: string;
  department?: string;
  position?: string;
}